Neatly add row_attributes to DefaultTable.
Move most logic into ::Base via include ::Extensions::RowAttributes.
Update Presenters::Default to add attributes to trs.
Add simple 'feature' spec in Presenters::Default spec.
