+++ lib/table_cloth/base.rb
@@ -1,6 +1,7 @@
 module TableCloth
   class Base
     include TableCloth::Extensions::Actions
+    include TableCloth::Extensions::RowAttributes
 
     NoPresenterError = Class.new(Exception)
 
@@ -58,21 +59,6 @@ module TableCloth
         else
           Configuration.new
         end
-      end
-
-      def row_attributes(*args, &block)
-        @tr_options ||= {}
-        options = args.extract_options! || {}
-        options[:proc] = block if block_given?
-        @tr_options = options
-      end
-
-      def tr_options
-        @tr_options ||= {}
-        if superclass.respond_to? :tr_options
-          @tr_options = superclass.tr_options.merge(@tr_options)
-        end
-        @tr_options
       end
     end
   end
