+++ spec/support/dummy_table.rb
@@ -1,4 +1,7 @@
 class DummyTable < TableCloth::Base
+  row_attributes name: "What a handsome quazimodo" do |obj|
+    {class: "quazimodo-is-awesome"} if obj.present?
+  end
   column :id, th_options: {class: "th_options_class"}
   column :name
   column :email
@@ -7,5 +10,5 @@ end
 class DummyTableUnlessAdmin < TableCloth::Base
   column :id, unless: :admin?
   column :name
-  column :email  
+  column :email
 end
